Move common_filter to lfp module
