+++ src/spyglass/common/__init__.py
@@ -14,7 +14,7 @@ from .common_ephys import (
     Raw,
     SampleCount,
 )
-from .common_filter import FirFilter
+from ..lfp.lfp_filter import FirFilter
 from .common_interval import (
     IntervalList,
     interval_list_censor,
